requirements.txt
asgiref==3.2.7
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
coverage==5.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.6
django-allauth==0.41.0
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-profiles==0.2
export==0.1.2
Faker==0.9.1
form==0.0.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
mixer==6.1.3
more-itertools==8.3.0
oauthlib==3.1.0
packaging==20.4
Pillow==7.1.2
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
pytest-cov==2.8.1
pytest-django==3.9.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
static-vars==0.1.0
text-unidecode==1.2
urllib3==1.25.9
wcwidth==0.1.9
whitenoise==5.1.0
zipp==3.1.0
